完成employee，studio模块，文件相对比较分散

+++ src/dao/StudioDAO.java
@@ -5,6 +5,7 @@ import idao.StudioIDAO;
 import java.sql.Connection;
 import java.sql.PreparedStatement;
 import java.sql.ResultSet;
+import java.sql.SQLException;
 import java.util.ArrayList;
 
 import model.StudioMODEL;
@@ -12,26 +13,115 @@ import mysql.ConnectionManager;
 
 public class StudioDAO implements StudioIDAO
 {
+
+    public static final int PAGE_SIZE = 8; // 每页显示条数
+    private int allCount; // 数据库中条数
+    private int allPageCount; // 总页数
+    private int currentPage; // 当前页
+
+    public int getAllCount()
+    {
+        return allCount;
+    }
+
+    public int getAllPageCount()
+    {
+        return allPageCount;
+    }
+
+    public int getCurrentPage()
+    {
+        return currentPage;
+    }
+
+    @SuppressWarnings("finally")
+    public ArrayList<StudioMODEL> findStudioByPage(int cPage, String studio_name)
+    {
+        currentPage = cPage;
+        ArrayList<StudioMODEL> list = new ArrayList<StudioMODEL>();
+
+        // 若未指定查找演出厅，则默认全查
+        if(null == studio_name || studio_name.equals("null"))
+        {
+            studio_name = "";
+        }
+
+        Connection conn = null;
+        PreparedStatement pstmt = null;
+        ResultSet rs = null;
+        try
+        {
+            // 获取记录总数
+            String sql1 = "select count(studio_id) as AllRecord from studio where studio_name like ?";
+            conn = ConnectionManager.getInstance().getConnection();
+            pstmt = conn.prepareStatement(sql1);
+            pstmt.setString(1, "%" + studio_name + "%");
+            rs = pstmt.executeQuery();
+            if(rs.next())
+                allCount = rs.getInt("AllRecord");
+            rs.close();
+            pstmt.close();
+
+            // 记算总页数
+            allPageCount = (allCount + PAGE_SIZE - 1) / PAGE_SIZE;
+
+            // 如果当前页数大于总页数，则赋值为总页数
+            if(allPageCount > 0 && currentPage > allPageCount)
+                currentPage = allPageCount;
+
+            // 获取第currentPage页数据
+            String sql2 = "select * from studio where studio_name like ? limit ?,?";
+            pstmt = conn.prepareStatement(sql2);
+            pstmt.setString(1, "%" + studio_name + "%");
+            pstmt.setInt(2, PAGE_SIZE * (currentPage - 1));
+            pstmt.setInt(3, PAGE_SIZE);
+            rs = pstmt.executeQuery();
+            StudioMODEL studio = null;
+            while(rs.next())
+            {
+                studio = new StudioMODEL();
+                studio.setStudio_id(rs.getInt("studio_id"));
+                studio.setStudio_name(rs.getString("studio_name"));
+                studio.setStudio_row_count(rs.getInt("studio_row_count"));
+                studio.setStudio_col_count(rs.getInt("studio_col_count"));
+                studio.setStudio_introduction(rs.getString("studio_introduction"));
+                studio.setStudio_flag(rs.getInt("studio_flag"));
+                // 将该用户信息插入列表
+                list.add(studio);
+            }
+        }
+        catch(SQLException e)
+        {
+            e.printStackTrace();
+        }
+        finally
+        {
+            ConnectionManager.close(rs, pstmt, conn);
+            return list;
+        }
+    }
+
     @Override
     public boolean insert(StudioMODEL studio)
     {
         boolean result = false;
         if(studio == null)
-        {
             return result;
-        }
+
+        // 获取Connection
         Connection con = ConnectionManager.getInstance().getConnection();
         PreparedStatement pstmt = null;
         try
         {
-            String sql = "insert into Studio(studio_name, studio_row_count, studio_col_count, studio_introduction, studio_flag) values(?,?,?,?,?)";
+            String sql = "insert into studio(studio_name, studio_row_count, studio_col_count, studio_introduction,studio_flag) values(?,?,?,?,?)";
             pstmt = con.prepareStatement(sql);
             pstmt.setString(1, studio.getStudio_name());
             pstmt.setInt(2, studio.getStudio_row_count());
             pstmt.setInt(3, studio.getStudio_col_count());
             pstmt.setString(4, studio.getStudio_introduction());
             pstmt.setInt(5, studio.getStudio_flag());
-            pstmt.executeLargeUpdate();
+
+            pstmt.executeUpdate();
             result = true;
         }
         catch(Exception e)
@@ -44,18 +134,20 @@ public class StudioDAO implements StudioIDAO
             ConnectionManager.close(null, pstmt, con);
             return result;
         }
-
     }
 
     @Override
     public boolean delete(int studio_id)
     {
         boolean result = false;
+        if(studio_id == 0)
+            return result;
+
         Connection con = ConnectionManager.getInstance().getConnection();
         PreparedStatement pstmt = null;
         try
         {
-
+            // 删除子某个用户
             String sql = "delete from studio where studio_id=?";
             pstmt = con.prepareStatement(sql);
             pstmt.setInt(1, studio_id);
@@ -70,6 +162,7 @@ public class StudioDAO implements StudioIDAO
         }
         finally
         {
+            // 关闭连接
             ConnectionManager.close(null, pstmt, con);
             return result;
         }
@@ -81,18 +174,19 @@ public class StudioDAO implements StudioIDAO
         boolean result = false;
         if(studio == null)
             return result;
+
         Connection con = ConnectionManager.getInstance().getConnection();
         PreparedStatement pstmt = null;
         try
         {
-            String sql = "update studio set studio_name=?,studio_row_count=?,studio_col_count=?,studio_introduction=?,studio_flag=?";
+            String sql = "update studio set studio_name=?,studio_row_count=?,studio_col_count=?,studio_introduction=?,studio_flag=? where studio_id=?";
             pstmt = con.prepareStatement(sql);
             pstmt.setString(1, studio.getStudio_name());
             pstmt.setInt(2, studio.getStudio_row_count());
             pstmt.setInt(3, studio.getStudio_col_count());
             pstmt.setString(4, studio.getStudio_introduction());
             pstmt.setInt(5, studio.getStudio_flag());
-
+            pstmt.setInt(6, studio.getStudio_id());
             pstmt.executeUpdate();
             result = true;
         }
@@ -112,100 +206,27 @@ public class StudioDAO implements StudioIDAO
     public ArrayList<StudioMODEL> findStudioAll()
     {
         ArrayList<StudioMODEL> list = new ArrayList<StudioMODEL>();
-        StudioMODEL info = null;
-
-        Connection con = ConnectionManager.getInstance().getConnection();
-        PreparedStatement pstmt = null;
-        ResultSet rs = null;
-        try
-        {
-            // 获取所有演出厅信息
-            pstmt = con.prepareStatement("select * from studio");
-            rs = pstmt.executeQuery();
-            while(rs.next())
-            {
-                info = new StudioMODEL();
-                info.setStudio_id(rs.getInt("studio_id"));
-                info.setStudio_name(rs.getString("studio_name"));
-                info.setStudio_row_count(rs.getInt("studio_row_count"));
-                info.setStudio_col_count(rs.getInt("studio_col_count"));
-                info.setStudio_introduction(rs.getString("studio_introduction"));
-                info.setStudio_flag(rs.getInt("studio_flag"));
-                // 加入列表
-                list.add(info);
-            }
-        }
-        catch(Exception e)
-        {
-            e.printStackTrace();
-        }
-        finally
-        {
-            ConnectionManager.close(rs, pstmt, con);
-            return list;
-        }
-    }
-
-    @Override
-    public ArrayList<StudioMODEL> findStudioById(String studio_name)
-    {
-        if(studio_name == null || studio_name.equals(""))
-            return null;
-
-        ArrayList<StudioMODEL> list = new ArrayList<StudioMODEL>();
-        StudioMODEL info = null;
-
-        Connection con = ConnectionManager.getInstance().getConnection();
-        PreparedStatement pstmt = null;
-        ResultSet rs = null;
-        try
-        {
-            // 获取所有演出厅信息:模糊查询
-            pstmt = con.prepareStatement("select * from studio where studio_name like ?");
-            pstmt.setString(1, "%" + studio_name + "%");// 拼接模糊查询串
-            rs = pstmt.executeQuery();
-            while(rs.next())
-            {
-                info = new StudioMODEL();
-                info.setStudio_id(rs.getInt("studio_id"));
-                info.setStudio_name(rs.getString("studio_name"));
-                info.setStudio_row_count(rs.getInt("studio_row_count"));
-                info.setStudio_col_count(rs.getInt("studio_col_count"));
-                info.setStudio_introduction(rs.getString("studio_introduction"));
-                info.setStudio_flag(rs.getInt("studio_flag"));
-
-            }
-        }
-        catch(Exception e)
-        {
-            e.printStackTrace();
-        }
-        finally
-        {
-            ConnectionManager.close(rs, pstmt, con);
-            return list;
-        }
-    }
-
-    @Override
-    public StudioMODEL findStudioByFlag(int flag)
-    {
-        StudioMODEL info = null;
+        StudioMODEL studio = null;
         Connection con = ConnectionManager.getInstance().getConnection();
         PreparedStatement pstmt = null;
         ResultSet rs = null;
         try
         {
             // 获取所有用户数据
-            pstmt = con.prepareStatement("select * from user where flag=?");
-            pstmt.setInt(1, flag);
+            pstmt = con.prepareStatement("select * from studio");
             rs = pstmt.executeQuery();
-            if(rs.next())
+            while(rs.next())
             {
-                // 如果有值的话再实例化
-                info = new StudioMODEL();
-                info.setStudio_flag(flag);
-                info.setStudio_id(rs.getInt("studio_id"));
+                studio = new StudioMODEL();
+
+                studio.setStudio_id(rs.getInt("studio_id"));
+                studio.setStudio_name(rs.getString("studio_name"));
+                studio.setStudio_row_count(rs.getInt("studio_row_count"));
+                studio.setStudio_col_count(rs.getInt("studio_col_count"));
+                studio.setStudio_introduction(rs.getString("studio_introduction"));
+                studio.setStudio_flag(rs.getInt("studio_flag"));
+                // 加入列表
+                list.add(studio);
             }
         }
         catch(Exception e)
@@ -215,7 +236,90 @@ public class StudioDAO implements StudioIDAO
         finally
         {
             ConnectionManager.close(rs, pstmt, con);
-            return info;
+            return list;
         }
     }
+
+    @Override
+    public ArrayList<StudioMODEL> findStudioByName(String studioName)
+    {
+        if(studioName == null || studioName.equals(""))
+            return null;
+
+        ArrayList<StudioMODEL> list = new ArrayList<StudioMODEL>();
+        StudioMODEL studio = null;
+
+        Connection con = ConnectionManager.getInstance().getConnection();
+        PreparedStatement pstmt = null;
+        ResultSet rs = null;
+        try
+        {
+            // 获取所有用户数据:模糊查询
+            pstmt = con.prepareStatement("select * from studio where studio_name like ?");
+            pstmt.setString(1, "%" + studioName + "%");// 拼接模糊查询串
+            rs = pstmt.executeQuery();
+            while(rs.next())
+            {
+                studio = new StudioMODEL();
+
+                studio.setStudio_id(rs.getInt("studio_id"));
+                studio.setStudio_name(rs.getString("studio_name"));
+                studio.setStudio_row_count(rs.getInt("studio_row_count"));
+                studio.setStudio_col_count(rs.getInt("studio_col_count"));
+                studio.setStudio_introduction(rs.getString("studio_introduction"));
+                studio.setStudio_flag(rs.getInt("studio_flag"));
+                // 加入列表
+                list.add(studio);
+            }
+        }
+        catch(Exception e)
+        {
+            e.printStackTrace();
+        }
+        finally
+        {
+            ConnectionManager.close(rs, pstmt, con);
+            return list;
+        }
+    }
+
+    @Override
+    public StudioMODEL findStudioById(int studio_id)
+    {
+        StudioMODEL studio = null;
+        if(studio_id == 0)
+            return studio;
+
+        Connection con = ConnectionManager.getInstance().getConnection();
+        PreparedStatement pstmt = null;
+        ResultSet rs = null;
+        try
+        {
+            // 获取所有用户数据
+            pstmt = con.prepareStatement("select * from studio where studio_id=?");
+            pstmt.setInt(1, studio_id);
+            rs = pstmt.executeQuery();
+            if(rs.next())
+            {
+                studio = new StudioMODEL();
+
+                studio.setStudio_id(rs.getInt("studio_id"));
+                studio.setStudio_name(rs.getString("studio_name"));
+                studio.setStudio_row_count(rs.getInt("studio_row_count"));
+                studio.setStudio_col_count(rs.getInt("studio_col_count"));
+                studio.setStudio_introduction(rs.getString("studio_introduction"));
+                studio.setStudio_flag(rs.getInt("studio_flag"));
+            }
+        }
+        catch(Exception e)
+        {
+            e.printStackTrace();
+        }
+        finally
+        {
+            ConnectionManager.close(rs, pstmt, con);
+            return studio;
+        }
+    }
+
 }
